+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+
+- Support parsing multiline doctests with `iex>` on all lines, but reformat them using `...>` on every line but the first one.
+
 ## 0.1.0 (2024-02-25)
 
 - Initial release.